Feat/mcp (#2)
* feat: add MCP Server and OpenClaw skill for VM-side GUI automation

- mcp/server/vsock_client.py: VM-side AF_VSOCK client using ClawShell
  FrameCodec (4B length prefix + JSON body), replacing the old binary
  frame protocol (magic + type + task_id in header). All ax operations
  route through call_capability() unified entry point.

- mcp/server/mcp_server.py: stdio JSON-RPC MCP Server exposing 7 GUI
  tools (list_windows, begin_task, end_task, click, get_ui_tree,
  set_value, activate_window). OpenClaw's acpx extension launches this
  as a subprocess and auto-discovers tools via get_manifest.

- mcp/client/clawshell-gui/SKILL.md: OpenClaw skill definition that
  guides the agent on when and how to use ClawShell GUI tools,
  including the typical workflow pattern.

- docs/vmm-progress.md: update progress, document mcp/ directory
  structure and OpenClaw integration architecture.

* feat: align MCP Server with standard protocol, add rootfs build script

- mcp_server.py: implement MCP standard methods (initialize,
  tools/list, tools/call), rename tool names from gui.xxx to gui__xxx,
  use inputSchema (camelCase), return content array in tools/call
  response. Server is now fully compatible with OpenClaw acpx stdio
  transport.

- build-rootfs.sh: WSL2 rootfs build script using debootstrap on
  Debian bookworm-slim. Installs Python3, Node.js 22, pnpm, OpenClaw,
  deploys ClawShell MCP Server to /opt/clawshell/mcp. MCP Server
  launched by acpx on demand, no systemd service.

- SKILL.md: sync tool names to gui__xxx format.

* fix: convert all script/Python files to Unix LF line endings

Add .gitattributes to enforce LF for *.sh and *.py, CRLF for *.sln/*.vcxproj.
Fixes 'bash\r: No such file or directory' when running build-rootfs.sh in WSL.

* fix: configure PNPM_HOME for pnpm global install in chroot

Set PNPM_HOME=/usr/local/share/pnpm in chroot stages 4/5/7 and
write /etc/profile.d/pnpm.sh so the env persists after rootfs boot.
Fixes ERR_PNPM_NO_GLOBAL_BIN_DIR during OpenClaw installation.

* feat: add checkpoint-based resume to rootfs build script

Each stage saves a tarball snapshot to /var/cache/clawshell-build/.
On re-run, completed stages are skipped automatically.

New options:
  --list     show checkpoint status
  --from N   rebuild from stage N
  --clean    wipe all checkpoints

* feat: add download caching to rootfs build for unreliable networks

- Split debootstrap into download + offline install using
  --make-tarball / --unpack-tarball, so deb packages are cached
  locally and network failure during download can be retried
  without re-downloading already fetched packages.

- Bind-mount a persistent apt cache dir into chroot so apt-get
  install stages also reuse previously downloaded debs.

- Add Acquire::Retries "3" apt config for automatic retry.

- New --purge-cache option to clear download cache.
- --clean only clears checkpoints, preserves download cache.

* fix: rootfs build - strict error checking and source dir detection

Three fixes:
- Stage 5: fail immediately if openclaw install fails, do NOT save
  checkpoint on failure (removed || echo that swallowed errors)
- Add git to debootstrap --include (openclaw dependency needs it)
- Add --src option and auto-detect ClawShell source dir from common
  WSL mount paths, with clear error message if not found

* feat: add comprehensive preflight check before rootfs build

Run all dependency checks upfront before starting any build stage:
- Root permission
- Host tools (debootstrap, tar, gzip, mount, chroot) with auto-install
- ClawShell source dir and required files validation
- Disk space on /tmp and cache partition
- Network connectivity to Debian mirror (warning only)

All errors are collected and reported together so the user
can fix everything in one pass instead of discovering issues
mid-build. --list and --purge-cache skip preflight since they
don't perform builds.

* fix: install git inside chroot for openclaw dependency

pnpm runs inside chroot which is isolated from the host.
@whiskeysockets/baileys needs 'git ls-remote' during install.
Also add build-essential and python3-dev for native addon builds.

* fix: extglob syntax error and output file location

- Replace bash extglob pattern !(en|en_US|locale.alias) with
  find -mindepth/-maxdepth which works in bash -c without shopt
- Change default output path to $CACHE_DIR/clawshell-rootfs.tar.gz
  (absolute path) so the file is always findable regardless of cwd
- Convert user-supplied output path to absolute
- Improve final instructions with copy-to-Windows step

* fix: output rootfs tarball to current directory by default

* refactor: redesign build-rootfs as a distributable tool

- Add --help with complete usage guide and stage descriptions
- Add --rebuild for clean + build in one command
- --clean now only cleans (checkpoint + download cache), no build
- Checkpoint: only keep one (latest stage), not 7 separate tarballs
- Auto-clean checkpoint after successful build to save disk
- Stage 7 skips checkpoint (packaging follows immediately)
- Add -o/--output option (short form)
- Cleaner command structure: ACTION-based dispatch

* refactor: remove --from option (redundant with single checkpoint)

* feat: add PowerShell installer for end-user deployment

install.ps1 provides a complete installation experience:
- Environment check (Windows version, WSL2 status)
- Download from GitHub Releases (BITS with resume support)
- Interactive config wizard (API provider + key)
- WSL distro import
- Write OpenClaw config into VM filesystem
- Register startup shortcut (user startup folder)
- Register in Add/Remove Programs for familiar uninstall
- Supports --Uninstall and --Upgrade modes

Usage: irm https://github.com/.../install.ps1 | iex

* feat: daemon-VMM integration, UI status pipeline, FrameCodec overlapped I/O

- Add VmmLauncher: daemon manages vmm.exe lifecycle (CreateProcess, graceful shutdown, monitor thread)
- Add VmmApp: vmm.exe refactored into application class with signal handling and clean shutdown
- Add health thread in daemon: periodic VM state and OpenClaw probe, pushes status to UI via Channel 2
- Wire Channel 3 (AF_HYPERV vsock) into daemon for direct VM-to-daemon communication
- Add FrameCodec::readFrameAsync / writeFrameAsync for OVERLAPPED Named Pipe I/O,
  fixing deadlock when serviceLoop read blocks concurrent push() writes on same handle
- Rework UIService to use FrameCodec async API with FILE_FLAG_OVERLAPPED pipe
- Extend UI (C# WinForms): StatusPanel with VM/OpenClaw/Channel indicators,
  AppState with full connection state tracking, MainForm status bar
- Update Channel 2 protocol: status message now carries vm/openclaw/channel fields
- Fix C4244 narrowing warning in ax_provider_windows.cc VK_ key map
- Update config, installer, build-rootfs script, and SKILL.md

+++ daemon-service/capability/ax/ax_provider_windows.cc
@@ -637,7 +637,7 @@ static void sendMouseClick(POINT pt, int button, int clicks)
 // 确保 MCP 工具侧的键名无需区分平台。
 static WORD keyNameToVk(const std::string& key)
 {
-	static const std::unordered_map<std::string, WORD> kMap = {
+	static const std::unordered_map<std::string, int> kMap = {
 		{"Return",   VK_RETURN},   {"Enter",    VK_RETURN},
 		{"Escape",   VK_ESCAPE},   {"Esc",      VK_ESCAPE},
 		{"Space",    VK_SPACE},
@@ -667,7 +667,7 @@ static WORD keyNameToVk(const std::string& key)
 	};
 	auto it = kMap.find(key);
 	if (it != kMap.end()) {
-		return it->second;
+		return static_cast<WORD>(it->second);
 	}
 	// 单字符：将 UTF-8 解码为宽字符后使用 VkKeyScanW。
 	// VkKeyScanW 对 ASCII 字符（≤0x7F）与 VkKeyScanA 等价；
